jdbc

+++ 4 Java/4_8 JDBC/JDBC/Code/JDBCApp/src/main/java/ro/teamnet/zth/api/em/EntityManager.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface EntityManager {
 
     <T> T findById(Class<T> entityClass, Long id);
-    int getNextIdVal(String tableName, String columnIdName);
+    long getNextIdVal(String tableName, String columnIdName);
     <T> Object insert(T entity);
     <T> List<T> findAll(Class<T> entityClass);
 }
